refactor for re-usable token mint input
