+++ components/TreasuryAccount/CreateAta.tsx
@@ -20,6 +20,8 @@ import useRealm from '@hooks/useRealm'
 import useGovernanceAssetsStore from 'stores/useGovernanceAssetsStore'
 import * as serum from '@project-serum/common'
 import { InformationCircleIcon } from '@heroicons/react/outline'
+import TokenMintInput from '@components/inputs/TokenMintInput'
+import { TokenInfo } from '@solana/spl-token-registry'
 
 const CreateAta = ({
   owner,
@@ -36,45 +38,16 @@ const CreateAta = ({
   )
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
-  const [isTyping, setIsTyping] = useState(false)
   const [isLoading, setIsLoading] = useState(false)
-  const [query, setQuery] = useState<string>('')
-  const [mintInfo, setMintInfo] = useState<
-    TokenProgramAccount<MintInfo> | undefined
-  >(undefined)
-  const tokenList = tokenService._tokenList
-  const foundByNameToken = tokenList.find(
-    (x) =>
-      x.address.toLowerCase() === query.toLowerCase() ||
-      x.name.toLowerCase() === query.toLowerCase() ||
-      x.symbol.toLowerCase() === query.toLowerCase()
-  )
-  const typedMint = tryParsePublicKey(query) ? query : ''
-  useEffect(() => {
-    const validateMint = async () => {
-      const info = await tryGetMint(
-        connection.current,
-        new PublicKey(typedMint)
-      )
-      setMintInfo(info)
-    }
-    if (typedMint) {
-      validateMint()
-    } else {
-      setMintInfo(undefined)
-    }
-  }, [typedMint])
-  useEffect(() => {
-    if (isTyping !== !!query) {
-      setIsTyping(!!query)
-    }
-    debounce.debounceFcn(async () => {
-      setIsTyping(false)
-    })
-  }, [query])
+  const [validatedTypedMint, setValidatedTypedMint] = useState<
+    string | undefined
+  >()
+  const [foundByNameToken, setFoundByNameToken] = useState<
+    TokenInfo | undefined
+  >()
   const handleCreate = async () => {
-    const mintPk = typedMint
-      ? new PublicKey(typedMint)
+    const mintPk = validatedTypedMint
+      ? new PublicKey(validatedTypedMint)
       : new PublicKey(foundByNameToken!.address)
     if (!mintPk) {
       throw 'Invalid mint'
@@ -127,78 +100,18 @@ const CreateAta = ({
   return (
     <div>
       <h3 className="mb-4 flex items-center">Create token account</h3>
-      <div>
-        <Input
-          noMaxWidth={true}
-          className="mb-2"
-          label="Token"
-          placeholder={'Mint, symbol or name'}
-          value={query}
-          type="text"
-          onChange={(evt) => setQuery(evt.target.value)}
-        />
-      </div>
-
-      <>
-        <div className="text-xs" style={{ minHeight: '16px' }}>
-          {!isTyping && (
-            <>
-              <div className="text-green">
-                {((typedMint && mintInfo) || foundByNameToken) && (
-                  <div>Token found</div>
-                )}
-              </div>
-              <div className="text-red">
-                {!mintInfo && !foundByNameToken && query && (
-                  <div>Token not found</div>
-                )}
-              </div>
-            </>
-          )}
-        </div>
-        <div className="mt-1">
-          <div
-            className="flex items-center text-fgd-1 border border-fgd-4 p-3 rounded-lg w-full"
-            style={{ minHeight: '60px' }}
-          >
-            {!isTyping && foundByNameToken ? (
-              <>
-                {foundByNameToken?.logoURI && (
-                  <img
-                    className={`flex-shrink-0 h-6 w-6 mr-2.5 mt-0.5`}
-                    src={foundByNameToken?.logoURI}
-                    onError={({ currentTarget }) => {
-                      currentTarget.onerror = null // prevents looping
-                      currentTarget.hidden = true
-                    }}
-                  />
-                )}
-                <div className="w-full">
-                  <div className="flex items-start justify-between mb-1">
-                    <div className="text-xs text-th-fgd-1">
-                      {foundByNameToken?.name}
-                    </div>
-                  </div>
-                  <div className="text-fgd-3 text-xs">
-                    {foundByNameToken?.symbol}
-                  </div>
-                </div>
-              </>
-            ) : (
-              <>
-                <span className="text-primary-light text-sm flex items-center">
-                  <InformationCircleIcon className="w-5 mr-1"></InformationCircleIcon>{' '}
-                  Type exact mint address, token name or symbol
-                </span>
-              </>
-            )}
-          </div>
-        </div>
-      </>
+      <TokenMintInput
+        onValidMintChange={(mintAddress, tokenInfo) => {
+          // TODO: Set the validated typedMint
+          setValidatedTypedMint(mintAddress)
+          // TODO: set the foundByNameToken
+          setFoundByNameToken(tokenInfo)
+        }}
+      />
       <div className="flex flex-col sm:flex-row sm:space-x-4 space-y-4 sm:space-y-0 mt-4">
         <Button
           className="ml-auto"
-          disabled={isLoading || (!typedMint && !foundByNameToken)}
+          disabled={isLoading || (!validatedTypedMint && !foundByNameToken)}
           onClick={handleCreate}
           isLoading={isLoading}
         >
